pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = ""
authors = ["Kazuki Kijima <kazukiyoshida96@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
poetry-plugin-export = "^1.8.0"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
pytest-django = "^4.8.0"
django-rest-framework = "^0.1.0"
httpx = "^0.27.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
redis = "^5.0.4"
pytest-mock = "^3.14.0"
django-cors-headers = "^4.3.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "dashboard.settings"
python_files = ["tests.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
